raspi_requirements.txt
mysql-connector-python==9.0.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
dnspython==2.6.1
mysql-connector-python==9.0.0
paho-mqtt==2.1.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-etcd==0.4.5
urllib3==2.2.2
dnspython==2.6.1
mysql-connector-python==9.0.0
paho-mqtt==2.1.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pyscard==2.0.10
python-etcd==0.4.5
urllib3==2.2.2
